delore-vscode-extension/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
captum==0.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
dgl==2.1.0
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gdown==5.1.0
gensim==4.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
libclang==18.1.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.30.1
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1
torchdata==0.7.1
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
